requirements.txt
addicty==2023.9.3
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asteval==1.0.8
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bibtexparser==1.4.4
blake3==1.0.8
bleach==6.3.0
bokeh==3.9.0
boto3==1.42.76
botocore==1.42.76
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
darkdetect==0.8.0
dask==2026.3.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
emmet-core==0.85.1
exceptiongroup==1.3.1
executing==2.2.1
fabio==2025.10.0
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
fsspec==2026.2.0
h11==0.16.0
h5py==3.16.0
hdf5plugin==6.0.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
ipykernel==7.2.0
ipython==8.38.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
larixite==2025.5.1
lark==1.3.1
latexcodec==3.0.1
lazy-loader==0.5
llvmlite==0.46.0
lmfit==1.3.4
locket==1.0.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdalib @ https://raw.githubusercontent.com/EPICS-synApps/utils/master/mdaPythonUtils/mdalib-2013.post2-py2.py3-none-any.whl
mdurl==0.1.2
mistune==3.2.0
monty==2025.3.3
mp-api==0.45.13
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.18.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.5
notebook_shim==0.2.4
numba==0.64.0
numdifftools==0.9.42
numexpr==2.14.1
numpy==2.2.6
orjson==3.11.7
overrides==7.7.0
packaging==26.0
palettable==3.3.3
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
partd==1.4.2
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py3Dmol==2.5.4
pyarrow==23.0.1
pybtex==0.25.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyFAI==2026.3.0
Pygments==2.19.2
pymatgen==2025.10.7
pymatgen-io-validation==0.1.2
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyparsing==3.3.2
pyshortcuts==1.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==2.1.0
sharrow==2.15.0
silx==2.2.2
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
sparse==0.17.0
spglib==2.7.0
SQLAlchemy==2.0.48
SQLAlchemy-Utils==0.42.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.10.0
termcolor==3.3.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tomli==2.4.1
tomli_w==1.2.0
toolz==1.1.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uncertainties==3.2.3
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
wrapt==2.1.2
wxmplot==2026.1.0
wxPython==4.2.5
wxutils==2026.1.0
xarray==2025.6.1
xmle==0.1.27
xraydb==4.5.8
xraylarch==2026.1.2
xyzservices==2025.11.0
yamllint==1.38.0
zipp==3.23.0
